# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Satcop Smart Visitor API",
    "description": "MVP FastAPI stub + Priority P2 Pickup (P1\u2013P6+H1) + after-hours (A1\u2013A6) + escort/zones Access Rules (B4) + Emergency Blast (E3 / B1\u2013B6). Contract \u00a7\u00a70\u20135 visit lifecycle unchanged. SMS mock / WA skipped_hold. Demo watermark; no production / live-school deploy.",
    "version": "0.6.0"
  },
  "paths": {
...
    "/v1/pickups": {
      "post": {
        "tags": [
          "pickups"
        ],
        "summary": "Start Pickup",
        "operationId": "start_pickup_v1_pickups_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickupCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "pickups"
        ],
        "summary": "List Pickups",
        "operationId": "list_pickups_v1_pickups_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dateFrom",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datefrom"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dateto"
            }
          },
          {
            "name": "studentId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Studentid"
            }
          },
          {
            "name": "gateId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gateid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "override",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Override"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pickups/{pickup_id}/consent": {
      "post": {
        "tags": [
          "pickups"
        ],
        "summary": "Record Consent",
        "operationId": "record_consent_v1_pickups__pickup_id__consent_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pickup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pickup Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickupConsentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pickups/{pickup_id}/release": {
      "post": {
        "tags": [
          "pickups"
        ],
        "summary": "Release Pickup",
        "operationId": "release_pickup_v1_pickups__pickup_id__release_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pickup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pickup Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickupReleaseBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pickups/{pickup_id}/request-override": {
      "post": {
        "tags": [
          "pickups"
        ],
        "summary": "Request Override",
        "operationId": "request_override_v1_pickups__pickup_id__request_override_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pickup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pickup Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pickups/{pickup_id}/override": {
      "post": {
        "tags": [
          "pickups"
        ],
        "summary": "Override Pickup",
        "operationId": "override_pickup_v1_pickups__pickup_id__override_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pickup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pickup Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pickups/{pickup_id}": {
      "get": {
        "tags": [
          "pickups"
        ],
        "summary": "Get Pickup",
        "operationId": "get_pickup_v1_pickups__pickup_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pickup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pickup Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdType": {
        "type": "string",
        "enum": [
          "Aadhaar",
          "DL",
          "Voter",
          "Passport",
          "Other"
        ],
        "title": "IdType"
      },
...
      "PickupConsentBody": {
        "properties": {
          "pickupConsentVersion": {
            "type": "string",
            "minLength": 1,
            "title": "Pickupconsentversion"
          },
          "pickupConsentAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickupconsentat"
          }
        },
        "type": "object",
        "required": [
          "pickupConsentVersion"
        ],
        "title": "PickupConsentBody"
      },
      "PickupCreate": {
        "properties": {
          "studentId": {
            "type": "string",
            "minLength": 1,
            "title": "Studentid"
          },
          "gateId": {
            "type": "string",
            "minLength": 1,
            "title": "Gateid"
          },
          "pickupReason": {
            "$ref": "#/components/schemas/PickupReason"
          },
          "reasonOther": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasonother"
          },
          "collectorPickupPersonId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collectorpickuppersonid"
          },
          "collectorName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collectorname"
          },
          "collectorMobile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collectormobile"
          },
          "idType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IdType"
              },
              {
                "type": "null"
              }
            ]
          },
          "idNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idnumber"
          },
          "idLast4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idlast4"
          },
          "linkVisit": {
            "type": "boolean",
            "title": "Linkvisit",
            "default": false
          },
          "pickupConsentAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickupconsentat"
          },
          "pickupConsentVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickupconsentversion"
          }
        },
        "type": "object",
        "required": [
          "studentId",
          "gateId",
          "pickupReason"
        ],
        "title": "PickupCreate"
      },
...
      "PickupReason": {
        "type": "string",
        "enum": [
          "early",
          "sick",
          "appointment",
          "other"
        ],
        "title": "PickupReason"
      },
...
      "PickupReleaseBody": {
        "properties": {
          "collectorLivePhotoRef": {
            "type": "string",
            "minLength": 1,
            "title": "Collectorlivephotoref"
          },
          "linkVisit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkvisit"
          }
        },
        "type": "object",
        "required": [
          "collectorLivePhotoRef"
        ],
        "title": "PickupReleaseBody"
      },
      "ReasonBody": {
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "ReasonBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}